requirements.txt
# Flask>=2.0
# python-telegram-bot==13.15
# requests

# Flask>=2.0
# requests
# python-dotenv
# python-dateutil
# pytz
# # Keep python-telegram-bot if you still use bot.py separately
# python-telegram-bot==13.15


# new code

Flask>=2.0
requests
python-dotenv
python-dateutil
pytz
Flask-Cors
Flask-Login
Flask-JWT-Extended
gunicorn
python-telegram-bot==13.15
pymongo # <<< ADDED for bson and MongoDB